backend/requirements.txt
# ===========================================
# ARIA Backend Dependencies - ALL FREE
# ===========================================

# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-dotenv==1.0.0

# AI - FREE (Groq + Local models)
groq==0.4.2
sentence-transformers==2.3.1
chromadb==0.4.22

# Database (Async only)
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1
pymongo==4.6.1
motor==3.3.2
redis==5.0.1

# Computer Vision - FREE (Updated versions)
opencv-python-headless
mediapipe
numpy
Pillow

# HTTP & APIs
httpx==0.26.0
aiohttp==3.9.3
websockets==12.0

# WhatsApp - Twilio
twilio==8.10.0

# Utilities
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Monitoring
sentry-sdk[fastapi]==1.39.2
structlog==24.1.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3